[move] itemmigrationパッケージの構造を整理する
