+++ src/main/scala/com/github/unchama/itemmigration/PlayerItemMigrationEntryPoints.scala
@@ -3,10 +3,10 @@ package com.github.unchama.itemmigration
 import cats.effect.{ConcurrentEffect, ContextShift, SyncEffect}
 import com.github.unchama.generic.ContextCoercion
 import com.github.unchama.generic.effect.unsafe.EffectEnvironment
-import com.github.unchama.itemmigration.controllers.player.{PlayerItemMigrationController, PlayerItemMigrationStateRepository}
+import com.github.unchama.itemmigration.bukkit.controllers.player.{PlayerItemMigrationController, PlayerItemMigrationStateRepository}
+import com.github.unchama.itemmigration.bukkit.targets.PlayerInventoriesData
 import com.github.unchama.itemmigration.domain.ItemMigrations
 import com.github.unchama.itemmigration.service.ItemMigrationService
-import com.github.unchama.itemmigration.targets.PlayerInventoriesData
 
 /**
  * プレーヤーのインベントリのマイグレーションを行うために必要なリスナー等のオブジェクトを提供するオブジェクトのクラス。
