+++ src/main/scala/com/github/unchama/itemmigration/bukkit/controllers/player/PlayerItemMigrationController.scala
@@ -1,11 +1,11 @@
-package com.github.unchama.itemmigration.controllers.player
+package com.github.unchama.itemmigration.bukkit.controllers.player
 
 import cats.effect.concurrent.TryableDeferred
 import cats.effect.{ConcurrentEffect, SyncEffect}
 import com.github.unchama.datarepository.bukkit.player.PreLoginToQuitPlayerDataRepository
+import com.github.unchama.itemmigration.bukkit.targets.PlayerInventoriesData
 import com.github.unchama.itemmigration.domain.ItemMigrations
 import com.github.unchama.itemmigration.service.ItemMigrationService
-import com.github.unchama.itemmigration.targets.PlayerInventoriesData
 import org.bukkit.entity.Player
 import org.bukkit.event.block.BlockPlaceEvent
 import org.bukkit.event.inventory.InventoryOpenEvent
