+++ src/main/scala/com/github/unchama/seichiassist/subsystems/itemmigration/infrastructure/loggers/PlayerItemsMigrationSlf4jLogger.scala
@@ -1,8 +1,8 @@
 package com.github.unchama.seichiassist.subsystems.itemmigration.infrastructure.loggers
 
 import cats.effect.Sync
+import com.github.unchama.itemmigration.bukkit.targets.PlayerInventoriesData
 import com.github.unchama.itemmigration.domain.{ItemMigrationLogger, ItemMigrationVersionNumber}
-import com.github.unchama.itemmigration.targets.PlayerInventoriesData
 import org.slf4j.Logger
 
 class PlayerItemsMigrationSlf4jLogger[F[_] : Sync](logger: Logger) extends ItemMigrationLogger[F, PlayerInventoriesData[F]] {
