+++ src/main/scala/com/github/unchama/seichiassist/subsystems/itemmigration/infrastructure/loggers/WorldLevelMigrationSlf4jLogger.scala
@@ -1,8 +1,8 @@
 package com.github.unchama.seichiassist.subsystems.itemmigration.infrastructure.loggers
 
 import cats.effect.Sync
+import com.github.unchama.itemmigration.bukkit.targets.WorldLevelData
 import com.github.unchama.itemmigration.domain.{ItemMigrationLogger, ItemMigrationVersionNumber}
-import com.github.unchama.itemmigration.targets.WorldLevelData
 import org.slf4j.Logger
 
 class WorldLevelMigrationSlf4jLogger[F[_]](logger: Logger)(implicit val F: Sync[F])
