+++ src/main/scala/com/github/unchama/seichiassist/subsystems/itemmigration/infrastructure/repositories/PlayerItemsMigrationVersionRepository.scala
@@ -1,8 +1,8 @@
 package com.github.unchama.seichiassist.subsystems.itemmigration.infrastructure.repositories
 
 import cats.effect.{Resource, Sync}
+import com.github.unchama.itemmigration.bukkit.targets.PlayerInventoriesData
 import com.github.unchama.itemmigration.domain.{ItemMigrationVersionNumber, ItemMigrationVersionRepository}
-import com.github.unchama.itemmigration.targets.PlayerInventoriesData
 import scalikejdbc._
 
 class PlayerItemsMigrationVersionRepository[F[_]](serverId: String)(implicit F: Sync[F])
